requirements.txt
mysql-connector-python
cryptography
pandas
openpyxl
Flask
python-dotenv
selenium
